Checkpoint work for importing new versions of crypto/openssl
This process is incredibly error prone today. This is an attempt to
scrape the build using gmake, then take the parsed information and emit
a templated Makefile using Jinja2.

This set of tools uses json as an intermediate format to make conversion
from one form to another more modular, and the Jinja2 template used in
`openssl_target_json_to_world_mk` is hardcoded in the script instead of
being a modular template.

Signed-off-by: Enji Cooper <yaneurabeya@gmail.com>